refactor BuildProfile
