+++ .gitignore
@@ -27,3 +27,6 @@ benchmarks/
 
 # Generated files in test directories
 sway-lsp/tests/**/Forc.lock
+
+# Generated files in example directories
+examples/**/Forc.lock
