+++ test/src/e2e_vm_tests/harness.rs
@@ -5,7 +5,7 @@ use forc_client::{
     op::{deploy, run},
     NodeTarget,
 };
-use forc_pkg::{manifest::ExperimentalFlags, BuildProfile, Built, BuiltPackage};
+use forc_pkg::{manifest::build_profile::ExperimentalFlags, BuildProfile, Built, BuiltPackage};
 use fuel_tx::TransactionBuilder;
 use fuel_vm::fuel_tx;
 use fuel_vm::interpreter::Interpreter;
